impl custom editor
renderer を pulldown で選択できるようになった
